Pull statuses into their own module
